security: Break dependency from v::config to v::security
Split v::security_config out from v::security.

v::config depends on v::security_config instead of v::security.

This will allow v::security to depend on v::http
