+++ src/v/cluster/CMakeLists.txt
@@ -197,6 +197,7 @@ v_cc_library(
     v::raft
     Roaring::roaring
     absl::flat_hash_map
+    v::security
     v::model
     v::cloud_storage
     v::features
