+++ src/v/config/CMakeLists.txt
@@ -13,9 +13,9 @@ v_cc_library(
   DEPS
     v::json
     v::model
-    v::security
-   boost_filesystem
-   absl::node_hash_set
-   yaml-cpp
+    v::security_config
+    boost_filesystem
+    absl::node_hash_set
+    yaml-cpp
 )
 add_subdirectory(tests)
