+++ src/v/security/gssapi_principal_mapper.cc
@@ -17,6 +17,7 @@
 #include <boost/algorithm/string/case_conv.hpp>
 #include <fmt/format.h>
 #include <fmt/ostream.h>
+#include <re2/re2.h>
 
 #include <charconv>
 
@@ -29,92 +30,17 @@ static std::string_view spv(const re2::StringPiece& sp) {
 }
 
 namespace security {
+namespace gssapi {
 static constexpr std::string_view gssapi_name_pattern
   = R"(([^/@]*)(/([^/@]*))?@([^/@]*))";
 static constexpr std::string_view non_simple_pattern = R"([/@])";
 static constexpr std::string_view parameter_pattern = R"(([^$]*)(\$(\d*))?)";
-static constexpr std::string_view rule_pattern
-  = R"(((DEFAULT)|((RULE:\[(\d*):([^\]]*)](\(([^)]*)\))?(s\/([^\/]*)\/([^\/]*)\/(g)?)?\/?(L|U)?))))";
 
 namespace detail {
 std::vector<gssapi_rule>
-parse_rules(const std::vector<ss::sstring>& unparsed_rules) {
-    static thread_local const re2::RE2 rule_parser(
-      rule_pattern, re2::RE2::Quiet);
-
-    vassert(
-      rule_parser.ok(),
-      "Failed to build rule pattern regex: {}",
-      rule_parser.error());
-
-    if (unparsed_rules.empty()) {
-        return {gssapi_rule()};
-    }
-
-    std::vector<gssapi_rule> rv;
-    re2::StringPiece default_;
-    re2::StringPiece num_components_str;
-    re2::StringPiece format;
-    re2::StringPiece match_regex;
-    re2::StringPiece from_pattern;
-    re2::StringPiece to_pattern;
-    re2::StringPiece repeat;
-    re2::StringPiece upper_lower;
-    for (const auto& rule : unparsed_rules) {
-        const re2::StringPiece rule_piece(rule.data(), rule.size());
-        if (!re2::RE2::FullMatch(
-              rule_piece,
-              rule_parser,
-              nullptr,
-              &default_,
-              nullptr,
-              nullptr,
-              &num_components_str,
-              &format,
-              nullptr,
-              &match_regex,
-              nullptr,
-              &from_pattern,
-              &to_pattern,
-              &repeat,
-              &upper_lower)) {
-            throw std::runtime_error("GSSAPI: Invalid rule: " + rule);
-        }
-        if (!default_.empty()) {
-            rv.emplace_back();
-        } else {
-            int num_components = std::numeric_limits<int>::max();
-            auto conv_rc = std::from_chars(
-              num_components_str.begin(),
-              num_components_str.end(),
-              num_components);
-            if (conv_rc.ec != std::errc()) {
-                throw std::runtime_error(
-                  "Invalid rule - Invalid value for number of components: "
-                  + num_components_str.as_string());
-            }
-            gssapi_rule::case_change_operation case_change
-              = gssapi_rule::case_change_operation::noop;
-
-            if (upper_lower == "L") {
-                case_change = gssapi_rule::case_change_operation::make_lower;
-            } else if (upper_lower == "U") {
-                case_change = gssapi_rule::case_change_operation::make_upper;
-            }
-            rv.emplace_back(
-              num_components,
-              spv(format),
-              spv(match_regex),
-              spv(from_pattern),
-              spv(to_pattern),
-              gssapi_rule::repeat{repeat == "g"},
-              case_change);
-        }
-    }
-
-    return rv;
+parse_rules(const std::vector<ss::sstring>& unparsed_rules);
 }
-} // namespace detail
+} // namespace gssapi
 
 gssapi_name::gssapi_name(
   std::string_view primary, std::string_view host_name, std::string_view realm)
@@ -128,7 +54,7 @@ gssapi_name::gssapi_name(
 
 std::optional<gssapi_name> gssapi_name::parse(std::string_view principal_name) {
     static thread_local const re2::RE2 gssapi_name_regex(
-      gssapi_name_pattern, re2::RE2::Quiet);
+      gssapi::gssapi_name_pattern, re2::RE2::Quiet);
     vassert(
       gssapi_name_regex.ok(),
       "Invalid name pattern: {}",
@@ -164,32 +90,10 @@ const ss::sstring& gssapi_name::host_name() const noexcept {
 }
 
 const ss::sstring& gssapi_name::realm() const noexcept { return _realm; }
-
-gssapi_rule::gssapi_rule(
-  int number_of_components,
-  std::string_view format,
-  std::string_view match,
-  std::string_view from_pattern,
-  std::string_view to_pattern,
-  repeat repeat_,
-  case_change_operation case_change)
-  : _is_default(false)
-  , _number_of_components(number_of_components)
-  , _format(format)
-  , _match(match)
-  , _from_pattern(std::regex{
-      from_pattern.begin(),
-      from_pattern.length(),
-      std::regex_constants::ECMAScript | std::regex_constants::optimize})
-  , _from_pattern_str(from_pattern)
-  , _to_pattern(to_pattern)
-  , _repeat(repeat_)
-  , _case_change(case_change) {}
-
 std::optional<ss::sstring> gssapi_rule::apply(
   std::string_view default_realm, std::vector<std::string_view> params) const {
     static thread_local const re2::RE2 non_simple_regex(
-      non_simple_pattern, re2::RE2::Quiet);
+      gssapi::non_simple_pattern, re2::RE2::Quiet);
     const re2::StringPiece match_piece(_match.data(), _match.size());
     const re2::RE2 match_regex(match_piece, re2::RE2::Quiet);
     vassert(
@@ -307,7 +211,7 @@ std::optional<ss::sstring> gssapi_rule::apply(
 std::optional<ss::sstring> gssapi_rule::replace_parameters(
   std::string_view format, std::vector<std::string_view> params) {
     static thread_local const re2::RE2 parameter_parser(
-      parameter_pattern, re2::RE2::Quiet);
+      gssapi::parameter_pattern, re2::RE2::Quiet);
     vassert(
       parameter_parser.ok(),
       "Invalid parameter pattern: {}",
@@ -383,9 +287,10 @@ ss::sstring gssapi_rule::replace_substitution(
 gssapi_principal_mapper::gssapi_principal_mapper(
   config::binding<std::vector<ss::sstring>> principal_to_local_rules_cb)
   : _principal_to_local_rules_binding(std::move(principal_to_local_rules_cb))
-  , _rules{detail::parse_rules(_principal_to_local_rules_binding())} {
+  , _rules{gssapi::detail::parse_rules(_principal_to_local_rules_binding())} {
     _principal_to_local_rules_binding.watch([this]() {
-        _rules = detail::parse_rules(_principal_to_local_rules_binding());
+        _rules = gssapi::detail::parse_rules(
+          _principal_to_local_rules_binding());
     });
 }
 
@@ -433,15 +338,6 @@ std::ostream& operator<<(std::ostream& os, const gssapi_principal_mapper& m) {
     return os;
 }
 
-std::optional<ss::sstring>
-validate_kerberos_mapping_rules(const std::vector<ss::sstring>& r) noexcept {
-    try {
-        detail::parse_rules(r);
-    } catch (const std::exception& e) {
-        return e.what();
-    }
-    return std::nullopt;
-}
 } // namespace security
 
 // explicit instantiations so as to avoid bringing in <fmt/ranges.h> in the
